Move PlayerProgressBar class to a separate h(c)pp files
